issue Pick a scooter solved
